move the new splash page to root
