+++ app/components/splash/Home.jsx
